\name{show}
\alias{show}
\alias{show,Populations}
\alias{show,Populations-method}
\docType{methods}

\title{Show Method for Populations Class} 

\description{
This method lists out the slots contained in a \code{Populations} object.
}

\usage{
\S4method{show}{Populations}(object)
}

\arguments{
\item{object}{Object returned from \code{\link{flowMeans}}}
}

\author{
Nima Aghaeepour <\email{naghaeep@bccrc.ca}>
}

\seealso{
\code{\link{flowMeans}}
}

\keyword{print}





