\name{summary}
\alias{summary}
\alias{summary,Populations}
\alias{summary,Populations-method}
\docType{methods}

\title{Summary Method for flowMeans Object}

\description{
This method prints out various characteristics of the populations found by flowMeans.
}

\usage{
\S4method{summary}{Populations}(object,...)

}

\arguments{
  \item{object}{Object returned from \code{\link{flowMeans}}.}
  \item{...}{Object returned from \code{\link{flowMeans}}.}
}

\details{
This method prints out various characteristics of the populations found by flowMeans.
}

\author{
Nima Aghaeepour <\email{naghaeep@bccrc.ca}>
}

\seealso{
\code{\link{flowMeans}}
}

\keyword{print}
